requirements.txt
aiogram~=2.25.1
pydub~=0.25.1
vosk~=0.3.45
pip~=23.2
attrs~=23.1.0
Pillow~=10.0.0
click~=8.1.6
certifi~=2023.5.7
idna~=3.4
urllib3~=2.0.3
setuptools~=68.0.0
requests~=2.31.0
python-dotenv~=1.0.0
numpy~=1.25.1
youtube_transcript_api~=0.6.1
g4f
art
ctransformers